refactor(application,shared,unittests): rename CityDto namespace and query/handler classes for clarity
- Moved CityDto from Shared.DTOs.Countries to Shared.DTOs.Cities namespace to better reflect its purpose.
- Updated all references and mappings to use AirlineBookingSystem.Shared.DTOs.Cities.CityDto.
- Renamed GetCityByCountryIdHandler/Query to GetCitiesByCountryIdHandler/Query for consistency and clarity.
- Renamed GetByIdHandler/Query (address) to GetAddressByIdHandler/Query for explicitness.
- Updated related unit tests and usages to match new class and namespace names.
- Improved using statements and mapping profiles to align with new structure.
