+++ AirlineBookingSystem.Application/Features/Addresses/Queries/ById/GetAddressByIdHandler.cs
@@ -5,10 +5,10 @@ using MediatR;
 
 namespace AirlineBookingSystem.Application.Features.Addresses.Queries.ById;
 
-public class GetByIdHandler(IAddressRepository addressRepository, IMapper mapper) 
-    : IRequestHandler<GetByIdQuery, AddressDto?>
+public class GetAddressByIdHandler(IAddressRepository addressRepository, IMapper mapper) 
+    : IRequestHandler<GetAddressByIdQuery, AddressDto?>
 {
-    public async Task<AddressDto?> Handle(GetByIdQuery request, CancellationToken cancellationToken)
+    public async Task<AddressDto?> Handle(GetAddressByIdQuery request, CancellationToken cancellationToken)
     {
         var address = await addressRepository.GetByIdAsync(request.Id);
         return mapper.Map<AddressDto>(address);
