+++ AirlineBookingSystem.Application/Features/Addresses/Queries/ById/GetAddressByIdQuery.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace AirlineBookingSystem.Application.Features.Addresses.Queries.ById;
 
-public record GetByIdQuery(int Id) : IRequest<AddressDto?>;
+public record GetAddressByIdQuery(int Id) : IRequest<AddressDto?>;
