+++ AirlineBookingSystem.Application/Features/Cities/Queries/ByCountryId/GetCitiesByCountryIdHandler.cs
@@ -1,14 +1,15 @@
 using AirlineBookingSystem.Application.Interfaces.Repositories;
+using AirlineBookingSystem.Shared.DTOs.Cities;
 using AirlineBookingSystem.Shared.DTOs.Countries;
 using AutoMapper;
 using MediatR;
 
 namespace AirlineBookingSystem.Application.Features.Cities.Queries.ByCountryId;
 
-public class GetCityByCountryIdHandler (ICityRepository cityRepository, IMapper mapper)
-    : IRequestHandler<GetCityByCountryIdQuery, IReadOnlyCollection<CityDto>>
+public class GetCitiesByCountryIdHandler (ICityRepository cityRepository, IMapper mapper)
+    : IRequestHandler<GetCitiesByCountryIdQuery, IReadOnlyCollection<CityDto>>
 {
-    public async Task<IReadOnlyCollection<CityDto>> Handle(GetCityByCountryIdQuery request, CancellationToken cancellationToken)
+    public async Task<IReadOnlyCollection<CityDto>> Handle(GetCitiesByCountryIdQuery request, CancellationToken cancellationToken)
     {
         var cities= cityRepository.GetByCountryIdAsync(request.CountryId);
         return mapper.Map<List<CityDto>>(await cities);
