+++ AirlineBookingSystem.Application/Features/Cities/Queries/ByCountryId/GetCitiesByCountryIdQuery.cs
@@ -1,6 +1,7 @@
+using AirlineBookingSystem.Shared.DTOs.Cities;
 using AirlineBookingSystem.Shared.DTOs.Countries;
 using MediatR;
 
 namespace AirlineBookingSystem.Application.Features.Cities.Queries.ByCountryId;
 
-public record GetCityByCountryIdQuery(int CountryId) : IRequest<IReadOnlyCollection<CityDto>>;
+public record GetCitiesByCountryIdQuery(int CountryId) : IRequest<IReadOnlyCollection<CityDto>>;
