+++ AirlineBookingSystem.Application/Features/Cities/Queries/ById/GetCityByIdQuery.cs
@@ -1,3 +1,4 @@
+using AirlineBookingSystem.Shared.DTOs.Cities;
 using AirlineBookingSystem.Shared.DTOs.Countries;
 using MediatR;
