+++ AirlineBookingSystem.Application/Mapping/CityProfile.cs
@@ -1,3 +1,4 @@
+using AirlineBookingSystem.Shared.DTOs.Cities;
 using AutoMapper;
 
 namespace AirlineBookingSystem.Application.Mapping;
@@ -6,7 +7,7 @@ public class CityProfile : Profile
 {
     public CityProfile()
     {
-        CreateMap<Domain.Entities.City, Shared.DTOs.Countries.CityDto>()
+        CreateMap<Domain.Entities.City, CityDto>()
             .ReverseMap();
     }
 }
