+++ AirlineBookingSystem.UnitTests/Cities/GetCityByIdHandlerTests.cs
@@ -1,6 +1,7 @@
 using AirlineBookingSystem.Application.Features.Cities.Queries.ById;
 using AirlineBookingSystem.Application.Interfaces.Repositories;
 using AirlineBookingSystem.Domain.Entities;
+using AirlineBookingSystem.Shared.DTOs.Cities;
 using AirlineBookingSystem.Shared.DTOs.Countries;
 using AutoMapper;
 using Moq;
